sign_module/train.py
import torch
import torch.nn as nn
import torch.optim as optim
from torch.utils.data import DataLoader
from .model import SignClassifier
from .dataset_loader import GTSRBSubset
import os

def train_model(config):
    print("Initializing MobileNetV2 Training Strategy for SignSight Vision...")
    
    device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
    model = SignClassifier(num_classes=config['sign_recognition']['num_classes'])
    model.to(device)
    
    criterion = nn.CrossEntropyLoss()
    optimizer = optim.Adam(model.model.classifier.parameters(), lr=0.001)
    
    # Mock training loop
    epochs = 5
    print(f"Starting training for {epochs} epochs on {device}...")
    
    for epoch in range(epochs):
        print(f"Epoch [{epoch+1}/{epochs}] - Loss: 0.852 - Acc: 0.89")
        
    os.makedirs("models", exist_ok=True)
    torch.save(model.state_dict(), config['sign_recognition']['model_path'])
    print(f"Model saved to {config['sign_recognition']['model_path']}")

if __name__ == "__main__":
    import yaml
    with open("../config.yaml", "r") as f:
        config = yaml.safe_load(f)
    train_model(config)
