requirements.txt
fastapi
uvicorn
jinja2
python-multipart
torch
torchvision
ultralytics
opencv-python
numpy
pyyaml
